move session expired component into components folder
