+++ embc-app/ClientApp/src/app/components/session-expired/session-expired.component.scss
