+++ embc-app/ClientApp/src/app/tester-page/tester-page.component.html
@@ -1,8 +1,5 @@
 <div class="row">
   <div class="col">
-    <pre>{{registrationSummary|json}}</pre>
-  </div>
-  <div class="col">
-    <pre>{{registration|json}}</pre>
+    Test Page
   </div>
 </div>
